+++ lib/src/bd_log_record.dart
@@ -7,13 +7,19 @@ class BDLogRecord {
   /// Create a new Logging record.
   /// Logging record's [level] is required.
   /// Logging record's [message] is required.
-  ///
   BDLogRecord(
     this.level,
     this.message, {
     this.error,
     this.stackTrace,
-  }) : time = DateTime.now();
+    this.isFatal = false,
+  })  : time = DateTime.now(),
+        assert(
+          (isFatal && error != null) ||
+              (!isFatal && error != null) ||
+              error == null,
+          'isFatal can only be used with error',
+        );
 
   /// Logging record's logging Level.
   final BDLevel level;
@@ -30,6 +36,9 @@ class BDLogRecord {
   /// Logging record's stacktrace.
   final StackTrace? stackTrace;
 
+  /// Logging record's error's fatal.
+  final bool isFatal;
+
   @override
   bool operator ==(Object other) =>
       identical(this, other) ||
@@ -39,7 +48,8 @@ class BDLogRecord {
           message == other.message &&
           error == other.error &&
           time == other.time &&
-          stackTrace == other.stackTrace;
+          stackTrace == other.stackTrace &&
+          isFatal == other.isFatal;
 
   @override
   int get hashCode =>
@@ -47,11 +57,12 @@ class BDLogRecord {
       message.hashCode ^
       error.hashCode ^
       time.hashCode ^
-      stackTrace.hashCode;
+      stackTrace.hashCode ^
+      isFatal.hashCode;
 
   @override
   String toString() {
     return 'LogRecord{level: $level, message: $message, error: $error, '
-        'time: $time, stackTrace: $stackTrace}';
+        'time: $time, stackTrace: $stackTrace, isFatal: $isFatal}';
   }
 }