+++ CHANGELOG.md
@@ -8,3 +8,9 @@
 ## 0.0.2
 
 * Decreased dependency boundaries on package collection.
+
+## 0.1.0
+
+* Added support for dart 3.0.0
+* Added Isolate file handler, that logs into a file but in another isolate.
+* Added a new BDLevel as success for success messages.